+++ lib/gettext/extractor.ex
@@ -318,6 +318,8 @@ defmodule Gettext.Extractor do
     %Translation{
       msgid: old.msgid,
       msgstr: old.msgstr,
+      # The new in-memory translation has no new flags.
+      flags: old.flags,
       # The new in-memory translation has no comments since it was extracted
       # from the source code.
       comments: old.comments,
@@ -356,6 +358,7 @@ defmodule Gettext.Extractor do
       msgid: old.msgid,
       msgid_plural: old.msgid_plural,
       msgstr: old.msgstr,
+      flags: old.flags,
       comments: old.comments,
       references: new.references
     }